# openapi.yaml
openapi: 3.1.0
info:
  title: Workload Governor API
  description: REST API for managing workload assignments on Stellar blockchain
  version: 1.0.0
  contact:
    name: Workload Governor Team
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.workload-governor.io
    description: Production server
paths:
...
  /api/v1/audit/logs:
    get:
      summary: Retrieve paginated audit log entries
      operationId: getAuditLogs
      tags:
        - Audit
      security:
        - BearerAuth: []
      parameters:
        - name: event_type
          in: query
          schema:
            type: string
            enum: [cap_update, maintainer_register, api_key_created, api_key_revoked]
          description: Filter by audit event type
        - name: actor
          in: query
          schema:
            type: string
          description: Filter by actor public key
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
          description: Number of log entries to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of log entries to skip
      responses:
        '200':
          description: Paginated list of audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogEntry'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                required:
                  - logs
                  - pagination
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Items skipped
        hasMore:
          type: boolean
          description: Whether more items exist
      required:
        - total
        - limit
        - offset
        - hasMore
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        correlationId:
          type: string
          nullable: true
          description: Correlation ID for tracing
      required:
        - error
...
    AuditLogEntry:
      type: object
      properties:
        id:
          type: integer
          description: Audit log entry ID
        event_type:
          type: string
          enum: [cap_update, maintainer_register, api_key_created, api_key_revoked]
          description: Type of admin event that was audited
        actor:
          type: string
          description: Public key or identifier of the admin who performed the action
        ip_address:
          type: string
          nullable: true
          description: IP address of the requester
        resource:
          type: string
          nullable: true
          description: The resource affected (e.g. "governance_caps")
        previous_value:
          nullable: true
          description: Serialised previous value before the change
        new_value:
          nullable: true
          description: Serialised new value after the change
        metadata:
          type: object
          nullable: true
          description: Additional context (e.g. correlation ID)
        created_at:
          type: string
          format: date-time
          description: Timestamp of the audit event
      required:
        - id
        - event_type
        - actor
        - created_at
  responses:
...
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: base64
      description: |
        Bearer token containing signed message in base64 format.
        Format: Bearer <base64(JSON.stringify({admin_address, message, signature}))>
...